Remove utils
No need for another library.
